refactor code

+++ src/main/java/com/google/graphgeckos/dashboard/fetchers/buildbot/BuildBotInitializer.java
@@ -2,9 +2,9 @@ package com.google.graphgeckos.dashboard.fetchers.buildbot;
 
 public class BuildBotInitializer {
   public final String name;
-  public final int initialBuildId;
+  public final long initialBuildId;
 
-  public BuildBotInitializer(String name, int initialBuildId) {
+  public BuildBotInitializer(String name, long initialBuildId) {
     this.name = name;
     this.initialBuildId = initialBuildId;
   }
